Backend/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-decouple==3.8
